Enabling the util-backend module & Cleanup
