+++ aide-processor/src/main/kotlin/team/duckie/quackquack/aide/processor/names.kt
@@ -7,11 +7,5 @@
 
 package team.duckie.quackquack.aide.processor
 
-internal const val QuackComponentPrefix = "Quack"
-
-internal const val ModifierSn = "Modifier"
-internal const val UnitSn = "Unit"
-
 internal const val ComposableFqn = "androidx.compose.runtime.Composable"
-
 internal const val DecorateModifierFqn = "team.duckie.quackquack.aide.annotation.DecorateModifier"
