+++ aide-processor/src/main/kotlin/team/duckie/quackquack/aide/processor/poet.kt
@@ -8,35 +8,22 @@
 package team.duckie.quackquack.aide.processor
 
 import com.google.devtools.ksp.processing.CodeGenerator
-import com.google.devtools.ksp.processing.Dependencies
 import com.google.devtools.ksp.processing.KSPLogger
 import com.google.devtools.ksp.symbol.KSFunctionDeclaration
-import com.squareup.kotlinpoet.AnnotationSpec
 import com.squareup.kotlinpoet.FileSpec
 import com.squareup.kotlinpoet.KModifier
 import com.squareup.kotlinpoet.ParameterizedTypeName.Companion.parameterizedBy
 import com.squareup.kotlinpoet.PropertySpec
 import com.squareup.kotlinpoet.asTypeName
 import com.squareup.kotlinpoet.buildCodeBlock
-import java.io.File
+import team.duckie.quackquack.util.backend.FormatterOffComment
+import team.duckie.quackquack.util.backend.SuppressAnnotation
+import team.duckie.quackquack.util.backend.generateBuildOrLocalFile
+import team.duckie.quackquack.util.backend.getGeneratedComment
+import team.duckie.quackquack.util.backend.requireContainingFile
+import team.duckie.quackquack.util.backend.toLiteralListString
 
-private const val GeneratedComment = "This file was automatically generated by aide-processor.\n" +
-        "Do not modify it arbitrarily."
-private const val FormatterOffComment = "\n@formatter:off"
-
-private val suppressAnnotation = AnnotationSpec
-    .builder(Suppress::class)
-    .addMember(
-        "%S, %S, %S, %S, %S, %S",
-        "NoConsecutiveBlankLines",
-        "PackageDirectoryMismatch",
-        "Wrapping",
-        "TrailingCommaOnCallSite",
-        "ArgumentListWrapping",
-        "ktlint",
-    )
-    .useSiteTarget(AnnotationSpec.UseSiteTarget.FILE)
-    .build()
+private val GeneratedComment = getGeneratedComment("aide-processor")
 
 private fun createQuackComponentsFileSpec(groupedComponents: List<Pair<String, Set<String>>>): FileSpec {
     val quackComponents = PropertySpec
@@ -76,7 +63,7 @@ private fun createQuackComponentsFileSpec(groupedComponents: List<Pair<String, S
         )
         .addFileComment(GeneratedComment)
         .addFileComment(FormatterOffComment)
-        .addAnnotation(suppressAnnotation)
+        .addAnnotation(SuppressAnnotation)
         .addProperty(quackComponents)
         .build()
 }
@@ -90,11 +77,12 @@ internal fun generateQuackComponents(
     val quackComponents = symbols.toDomainWithSimpleNameGroups()
     val quackComponentsFileSpec = createQuackComponentsFileSpec(quackComponents)
 
-    generateFile(
+    generateBuildOrLocalFile(
         codeGenerator = codeGenerator,
         fileSpec = quackComponentsFileSpec,
         path = aidePath,
         logger = logger,
+        tag = "aide",
     )
 }
 
@@ -137,7 +125,7 @@ private fun createAideModifiersFileSpec(groupedModifiers: List<Pair<String, Set<
         )
         .addFileComment(GeneratedComment)
         .addFileComment(FormatterOffComment)
-        .addAnnotation(suppressAnnotation)
+        .addAnnotation(SuppressAnnotation)
         .addProperty(aideModifiers)
         .build()
 }
@@ -151,40 +139,13 @@ internal fun generateAideModifiers(
     val aideModifiers = symbols.toDomainWithSimpleNameGroups()
     val aideModifiersFileSpec = createAideModifiersFileSpec(aideModifiers)
 
-    generateFile(
+    generateBuildOrLocalFile(
         codeGenerator = codeGenerator,
         fileSpec = aideModifiersFileSpec,
         path = aidePath,
         logger = logger,
+        tag = "aide",
     )
-}
-
-private fun generateFile(
-    codeGenerator: CodeGenerator,
-    fileSpec: FileSpec,
-    path: String?,
-    logger: KSPLogger,
-) {
-    val generatedPath: String
-    if (path == null) {
-        val fileOutstream = codeGenerator.createNewFile(
-            dependencies = Dependencies.Empty,
-            packageName = fileSpec.packageName,
-            fileName = fileSpec.name,
-        )
-        fileOutstream.writer().use(fileSpec::writeTo)
-        generatedPath = fileSpec.packageName + "/" + fileSpec.name + ".kt"
-    } else {
-        val file = File(path, "${fileSpec.name}.kt").also { file ->
-            if (!file.exists()) {
-                file.parentFile.mkdirs()
-                file.createNewFile()
-            }
-        }
-        file.writeText(fileSpec.toString())
-        generatedPath = file.path
-    }
-    logger.warn("[AIDE] generated at $generatedPath")
 }
 
 private fun Sequence<KSFunctionDeclaration>.toDomainWithSimpleNameGroups(): List<Pair<String, Set<String>>> {
