+++ aide/src/main/kotlin/team/duckie/quackquack/aide/rule/AideModifiers.kt
@@ -1,8 +1,9 @@
 // This file was automatically generated by aide-processor.
-// Do not modify it arbitrarily.
+// Do not modify it manually.
 // @formatter:off
 @file:Suppress("NoConsecutiveBlankLines", "PackageDirectoryMismatch", "Wrapping",
-    "TrailingCommaOnCallSite", "ArgumentListWrapping", "ktlint")
+    "TrailingCommaOnCallSite", "ArgumentListWrapping", "RedundantVisibilityModifier",
+    "UnusedImport", "NoUnusedImports", "SpacingAroundParens", "ktlint")
 
 import kotlin.String
 import kotlin.Suppress
