+++ build.gradle.kts
@@ -102,6 +102,7 @@ allprojects {
                 freeCompilerArgs = freeCompilerArgs + listOf(
                     "-opt-in=kotlin.OptIn",
                     "-opt-in=kotlin.RequiresOptIn",
+                    "-Xcontext-receivers",
                 )
             }
         }
