+++ casa-processor/src/main/kotlin/team/duckie/quackquack/casa/processor/names.kt
@@ -1,0 +1,16 @@
+/*
+ * Designed and developed by Duckie Team 2023.
+ *
+ * Licensed under the MIT.
+ * Please see full license: https://github.com/duckie-team/quack-quack-android/blob/main/LICENSE
+ */
+
+package team.duckie.quackquack.casa.processor
+
+import com.squareup.kotlinpoet.ClassName
+
+internal const val SugarReferSn = "SugarRefer"
+
+internal const val CasaValueSn = "CasaValue"
+internal val CasaModelCn = ClassName("team.duckie.quackquack.casa.material", "CasaModel")
+internal const val CasaFqn = "team.duckie.quackquack.casa.annotation.Casa"
