+++ catalog/src/main/kotlin/team/duckie/quackquack/catalog/CasaModels.kt
@@ -1,73 +1,17 @@
 // This file was automatically generated by casa-processor.
-// Do not modify it arbitrarily.
+// Do not modify it manually.
 // @formatter:off
 @file:Suppress("NoConsecutiveBlankLines", "PackageDirectoryMismatch", "Wrapping",
     "TrailingCommaOnCallSite", "ArgumentListWrapping", "RedundantVisibilityModifier",
     "UnusedImport", "NoUnusedImports", "SpacingAroundParens", "ktlint")
 
 import androidx.compose.runtime.Composable
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.text.style.TextOverflow
-import kotlin.Boolean
-import kotlin.String
 import kotlin.Suppress
 import kotlinx.collections.immutable.ImmutableList
 import kotlinx.collections.immutable.persistentListOf
 import kotlinx.collections.immutable.toImmutableList
 import team.duckie.quackquack.casa.material.CasaModel
-import team.duckie.quackquack.ui.sugar.QuackBody1
-import team.duckie.quackquack.ui.sugar.QuackBody2
-import team.duckie.quackquack.ui.sugar.QuackBody3
-import team.duckie.quackquack.ui.sugar.QuackHeadLine1
-import team.duckie.quackquack.ui.sugar.QuackHeadLine2
-import team.duckie.quackquack.ui.sugar.QuackLarge1
-import team.duckie.quackquack.ui.sugar.QuackSubtitle
-import team.duckie.quackquack.ui.sugar.QuackSubtitle2
-import team.duckie.quackquack.ui.sugar.QuackTitle1
-import team.duckie.quackquack.ui.sugar.QuackTitle2
-
-private val textQuackTextCasaModel: CasaModel = CasaModel(
-  name = "QuackText",
-  domain = "text",
-  kdocDefaultSection = """
-      |텍스트를 그리는 기본적인 컴포저블입니다.
-      |""".trimMargin(),
-  components = persistentListOf<Pair<String, @Composable () -> Unit>>(
-      "QuackBody1" to { QuackBody1(
-          text = "QuackText",
-      ) },
-      "QuackBody2" to { QuackBody2(
-          text = "QuackText",
-      ) },
-      "QuackBody3" to { QuackBody3(
-          text = "QuackText",
-      ) },
-      "QuackHeadLine1" to { QuackHeadLine1(
-          text = "QuackText",
-      ) },
-      "QuackHeadLine2" to { QuackHeadLine2(
-          text = "QuackText",
-      ) },
-      "QuackLarge1" to { QuackLarge1(
-          text = "QuackText",
-      ) },
-      "QuackSubtitle" to { QuackSubtitle(
-          text = "QuackText",
-      ) },
-      "QuackSubtitle2" to { QuackSubtitle2(
-          text = "QuackText",
-      ) },
-      "QuackTitle1" to { QuackTitle1(
-          text = "QuackText",
-      ) },
-      "QuackTitle2" to { QuackTitle2(
-          text = "QuackText",
-      ) },
-      ).toImmutableList(),
-)
-
 
 public val casaModels: ImmutableList<CasaModel> = persistentListOf(
-  textQuackTextCasaModel,
 )
