+++ sugar-processor/src/main/kotlin/team/duckie/quackquack/sugar/processor/ir/SugarIrVisitor.kt
@@ -34,8 +34,10 @@ import org.jetbrains.kotlin.kdoc.psi.impl.KDocTag
 import org.jetbrains.kotlin.name.FqName
 import org.jetbrains.kotlin.utils.addToStdlib.UnsafeCastFunction
 import org.jetbrains.kotlin.utils.addToStdlib.cast
-import team.duckie.quackquack.sugar.processor.Logger
-import team.duckie.quackquack.sugar.processor.prependLogPrefix
+import team.duckie.quackquack.util.backend.Logger
+import team.duckie.quackquack.util.backend.isQuackComponent
+import team.duckie.quackquack.util.backend.locationOf
+import team.duckie.quackquack.util.backend.prependLogPrefix
 
 internal class SugarIrVisitor(
     @Suppress("unused") private val context: IrPluginContext,
@@ -97,7 +99,7 @@ internal class SugarIrVisitor(
                 parameters = sugarParameters,
             )
 
-            logger(sugarIrData.prependLogPrefix(withNewline = true))
+            logger(with(logger) { sugarIrData.prependLogPrefix(withNewline = true) })
             addSugarIrData(sugarIrData)
         }
     }
@@ -114,11 +116,11 @@ private fun IrConstructorCall.getSugarNameIfNotDefault(): String? {
 }
 
 private fun checkCustomSugarNameIsValid(name: String) {
-    if (!name.startsWith(QuackComponentPrefix)) {
-        throw IllegalArgumentException(SourceError.sugarNamePrefixIsNotQuack(name))
+    require(name.startsWith(QuackComponentPrefix)) {
+        SourceError.sugarNamePrefixIsNotQuack(name)
     }
-    if (!name.contains(SugarTokenName)) {
-        throw IllegalArgumentException(SourceError.sugarNameWithoutTokenName(name))
+    require(name.contains(SugarTokenName)) {
+        SourceError.sugarNameWithoutTokenName(name)
     }
 }
 
@@ -133,7 +135,7 @@ private fun IrValueParameter.toSugarParameter(isToken: Boolean): SugarParameter 
         val sugarImportsExpression = sugarImportsAnnotation.getValueArgument(0)
         sugarImportsExpression.cast<IrVararg>().elements.map { element ->
             element.cast<IrClassReference>().classType.classFqName ?: error(
-                SourceError.importClazzFqnUnavailable(element),
+                message = SourceError.importClazzFqnUnavailable(element),
             )
         }
     }
@@ -148,14 +150,14 @@ private fun IrValueParameter.toSugarParameter(isToken: Boolean): SugarParameter 
     )
 }
 
+// TODO(2): 사용된 토큰 기재
+// TODO(1): util-backend로 KDoc 추출 공통 로직 분리
 private fun IrSimpleFunction.getSugarKDoc(referFqn: FqName, sugarTokenName: String): String {
     val generatedDocComment = "This document was automatically generated by [${referFqn.asString()}].\n" +
             "If any links are broken, please check the original document."
 
     val kdocArea = psiElement?.children?.firstOrNull { it is KDoc } as? KDoc
-    val kdocDefaultSection = kdocArea?.getDefaultSection()
-
-    if (kdocDefaultSection?.getContent() == null) return ""
+    val kdocDefaultSection = kdocArea?.getDefaultSection() ?: return ""
 
     val kdocTags = kdocArea.children
         .firstOrNull { it is KDocSection }
