+++ sugar-processor/src/main/kotlin/team/duckie/quackquack/sugar/processor/poet/PoetUtils.kt
@@ -7,20 +7,9 @@
 
 package team.duckie.quackquack.sugar.processor.poet
 
-import com.squareup.kotlinpoet.FileSpec
-import com.squareup.kotlinpoet.FunSpec
-import org.jetbrains.kotlin.name.FqName
 import team.duckie.quackquack.sugar.processor.ir.QuackComponentPrefix
 import team.duckie.quackquack.sugar.processor.ir.SugarIrData
 import team.duckie.quackquack.sugar.processor.ir.SugarTokenName
-
-// TODO: Testing
-internal fun String.bestGuessToKotlinPackageName(): String {
-    if (contains("src/main/java/")) {
-        error("The given package is a Java package.")
-    }
-    return substringAfter("src/main/kotlin/").replace("/", ".")
-}
 
 // TODO: Testing
 internal fun SugarIrData.toSugarComponentName(tokenFqExpression: String): String {
@@ -34,23 +23,4 @@ internal fun SugarIrData.toSugarComponentName(tokenFqExpression: String): String
             .toMutableList()
             .apply { addAll(QuackComponentPrefix.length, tokenExpression.toList()) }
             .joinToString("")
-}
-
-internal fun FileSpec.Builder.addImports(imports: List<FqName>): FileSpec.Builder {
-    return apply {
-        imports.forEach { import ->
-            addImport(
-                packageName = import.parent().asString(),
-                import.shortName().asString(),
-            )
-        }
-    }
-}
-
-internal fun FileSpec.Builder.addFunctions(functions: List<FunSpec>): FileSpec.Builder {
-    return apply {
-        functions.forEach { function ->
-            addFunction(function)
-        }
-    }
 }
