+++ sugar-processor/src/main/kotlin/team/duckie/quackquack/sugar/processor/poet/poet.kt
@@ -26,29 +26,17 @@ import team.duckie.quackquack.sugar.processor.ir.SugarGeneratorUsageCn
 import team.duckie.quackquack.sugar.processor.ir.SugarIrData
 import team.duckie.quackquack.sugar.processor.ir.SugarParameter
 import team.duckie.quackquack.sugar.processor.ir.SugarReferCn
+import team.duckie.quackquack.util.backend.FormatterOffComment
+import team.duckie.quackquack.util.backend.SuppressAnnotation
+import team.duckie.quackquack.util.backend.addFunctions
+import team.duckie.quackquack.util.backend.addImports
+import team.duckie.quackquack.util.backend.bestGuessToKotlinPackageName
+import team.duckie.quackquack.util.backend.getGeneratedComment
 
-private const val GeneratedComment = "This file was automatically generated by sugar-processor.\n" +
-        "Do not modify it arbitrarily."
-private const val FormatterOffComment = "\n@formatter:off"
-
-private val suppressAnnotation = AnnotationSpec
-    .builder(Suppress::class)
-    .addMember(
-        "%S, %S, %S, %S, %S, %S, %S, %S",
-        "NoConsecutiveBlankLines",
-        "ModifierParameter",
-        "RedundantUnitReturnType",
-        "Wrapping",
-        "TrailingCommaOnCallSite",
-        "NoUnitReturn",
-        "ArgumentListWrapping",
-        "ktlint",
-    )
-    .useSiteTarget(AnnotationSpec.UseSiteTarget.FILE)
-    .build()
+private val GeneratedComment = getGeneratedComment("sugar-processor")
 
 @Suppress("OPT_IN_CAN_ONLY_BE_USED_AS_ANNOTATION")
-private val sugarCompilerOptInAnnotation = AnnotationSpec
+private val SugarCompilerOptInAnnotation = AnnotationSpec
     .builder(OptIn::class)
     .addMember(
         "%T::class, %T::class",
@@ -58,7 +46,7 @@ private val sugarCompilerOptInAnnotation = AnnotationSpec
     .useSiteTarget(AnnotationSpec.UseSiteTarget.FILE)
     .build()
 
-private val sugarGeneratedFileMarkerAnnotation = AnnotationSpec
+private val SugarGeneratedFileMarkerAnnotation = AnnotationSpec
     .builder(SugarGeneratedFileCn)
     .useSiteTarget(AnnotationSpec.UseSiteTarget.FILE)
     .build()
@@ -80,9 +68,9 @@ internal fun generateSugarComponentFiles(
             )
             .addFileComment(GeneratedComment)
             .addFileComment(FormatterOffComment)
-            .addAnnotation(suppressAnnotation)
-            .addAnnotation(sugarCompilerOptInAnnotation)
-            .addAnnotation(sugarGeneratedFileMarkerAnnotation)
+            .addAnnotation(SuppressAnnotation)
+            .addAnnotation(SugarCompilerOptInAnnotation)
+            .addAnnotation(SugarGeneratedFileMarkerAnnotation)
             .addImports(imports.toMutableList().apply { add(SugarFqn) })
             .addFunctions(funSpecs)
             .build()
@@ -129,6 +117,7 @@ private fun SugarIrData.toFunSpecWithImports(tokenFqExpression: String): Pair<Li
 
                 val parameterName = parameter.name.asString()
                 val parameterValue = if (parameter.isToken) tokenFqExpression else parameterName
+
                 addStatement("%L = %L,", parameterName, parameterValue)
             }
         }
