+++ ui/report/compose-metrics/ui_debug-module.json
@@ -7,19 +7,19 @@
  "totalGroups": 23,
  "staticArguments": 14,
  "certainArguments": 73,
- "knownStableArguments": 142,
- "knownUnstableArguments": 11,
+ "knownStableArguments": 143,
+ "knownUnstableArguments": 12,
  "unknownStableArguments": 0,
- "totalArguments": 153,
+ "totalArguments": 155,
  "markedStableClasses": 4,
  "inferredStableClasses": 0,
  "inferredUnstableClasses": 0,
  "inferredUncertainClasses": 0,
  "effectivelyStableClasses": 4,
  "totalClasses": 4,
- "memoizedLambdas": 2,
+ "memoizedLambdas": 3,
  "singletonLambdas": 0,
  "singletonComposableLambdas": 0,
  "composableLambdas": 2,
- "totalLambdas": 5
+ "totalLambdas": 6
 }
