+++ ui/report/compose-metrics/ui_debugAndroidTest-module.json
@@ -1,25 +1,25 @@
 {
- "skippableComposables": 5,
- "restartableComposables": 8,
+ "skippableComposables": 6,
+ "restartableComposables": 9,
  "readonlyComposables": 0,
- "totalComposables": 8,
- "restartGroups": 8,
- "totalGroups": 8,
- "staticArguments": 11,
+ "totalComposables": 9,
+ "restartGroups": 9,
+ "totalGroups": 9,
+ "staticArguments": 14,
  "certainArguments": 0,
- "knownStableArguments": 64,
+ "knownStableArguments": 70,
  "knownUnstableArguments": 0,
  "unknownStableArguments": 0,
- "totalArguments": 64,
+ "totalArguments": 70,
  "markedStableClasses": 0,
  "inferredStableClasses": 0,
- "inferredUnstableClasses": 1,
+ "inferredUnstableClasses": 2,
  "inferredUncertainClasses": 0,
  "effectivelyStableClasses": 0,
- "totalClasses": 1,
- "memoizedLambdas": 5,
+ "totalClasses": 2,
+ "memoizedLambdas": 6,
  "singletonLambdas": 0,
  "singletonComposableLambdas": 1,
- "composableLambdas": 5,
- "totalLambdas": 5
+ "composableLambdas": 6,
+ "totalLambdas": 6
 }
