+++ ui/report/compose-metrics/ui_release-module.json
@@ -7,19 +7,19 @@
  "totalGroups": 23,
  "staticArguments": 14,
  "certainArguments": 73,
- "knownStableArguments": 141,
- "knownUnstableArguments": 10,
+ "knownStableArguments": 143,
+ "knownUnstableArguments": 12,
  "unknownStableArguments": 0,
- "totalArguments": 151,
+ "totalArguments": 155,
  "markedStableClasses": 4,
  "inferredStableClasses": 0,
  "inferredUnstableClasses": 0,
  "inferredUncertainClasses": 0,
  "effectivelyStableClasses": 4,
  "totalClasses": 4,
- "memoizedLambdas": 1,
+ "memoizedLambdas": 3,
  "singletonLambdas": 0,
  "singletonComposableLambdas": 0,
  "composableLambdas": 2,
- "totalLambdas": 4
+ "totalLambdas": 6
 }
