+++ ui/report/compose-reports/ui_debugAndroidTest-classes.txt
@@ -1,3 +1,7 @@
+unstable class ButtonTest {
+  unstable val rule: ComposeContentTestRule
+  <runtime stability> = Unstable
+}
 unstable class TextTest {
   unstable val rule: ComposeContentTestRule
   <runtime stability> = Unstable
