+++ ui/report/compose-reports/ui_release-composables.csv
@@ -1,5 +1,5 @@
 package,name,composable,skippable,restartable,readonly,inline,isLambda,hasDefaults,defaultsGroup,groups,calls,
-team.duckie.quackquack.ui.QuackButton,QuackButton,1,0,0,0,0,0,0,0,1,5,
+team.duckie.quackquack.ui.QuackButton,QuackButton,1,0,0,0,0,0,0,0,1,6,
 team.duckie.quackquack.ui.QuackBaseButton,QuackBaseButton,1,1,1,0,0,0,0,0,1,2,
 team.duckie.quackquack.ui.sugar.QuackBody1,QuackBody1,1,1,1,0,0,0,0,0,1,1,
 team.duckie.quackquack.ui.sugar.QuackBody2,QuackBody2,1,1,1,0,0,0,0,0,1,1,
@@ -11,7 +11,7 @@ team.duckie.quackquack.ui.sugar.QuackSubtitle,QuackSubtitle,1,1,1,0,0,0,0,0,1,1,
 team.duckie.quackquack.ui.sugar.QuackSubtitle2,QuackSubtitle2,1,1,1,0,0,0,0,0,1,1,
 team.duckie.quackquack.ui.sugar.QuackTitle1,QuackTitle1,1,1,1,0,0,0,0,0,1,1,
 team.duckie.quackquack.ui.sugar.QuackTitle2,QuackTitle2,1,1,1,0,0,0,0,0,1,1,
-team.duckie.quackquack.ui.QuackText,QuackText,1,1,1,0,0,0,0,0,4,6,
+team.duckie.quackquack.ui.QuackText,QuackText,1,1,1,0,0,0,0,0,4,7,
 team.duckie.quackquack.ui.QuackClickableText,QuackClickableText,1,1,1,0,0,0,0,0,1,3,
 team.duckie.quackquack.ui.rememberSpanAnnotatedString,rememberSpanAnnotatedString,1,0,0,0,0,0,0,0,1,1,
 team.duckie.quackquack.ui.util.currentFontScale,currentFontScale,1,0,0,1,1,0,0,0,1,1,
