+++ ui/src/main/kotlin/team/duckie/quackquack/ui/sugar/text.kt
@@ -1,8 +1,9 @@
 // This file was automatically generated by sugar-processor.
-// Do not modify it arbitrarily.
+// Do not modify it manually.
 // @formatter:off
-@file:Suppress("NoConsecutiveBlankLines", "ModifierParameter", "RedundantUnitReturnType",
-    "Wrapping", "TrailingCommaOnCallSite", "NoUnitReturn", "ArgumentListWrapping", "ktlint")
+@file:Suppress("NoConsecutiveBlankLines", "PackageDirectoryMismatch", "Wrapping",
+    "TrailingCommaOnCallSite", "ArgumentListWrapping", "RedundantVisibilityModifier",
+    "UnusedImport", "NoUnusedImports", "SpacingAroundParens", "ktlint")
 @file:OptIn(SugarCompilerApi::class, SugarGeneratorUsage::class)
 @file:SugarGeneratedFile
