+++ util-backend/build.gradle.kts
@@ -5,8 +5,19 @@
  * Please see full license: https://github.com/duckie-team/quack-quack-android/blob/main/LICENSE
  */
 
+@file:Suppress("INLINE_FROM_HIGHER_PLATFORM")
+
 plugins {
     quackquack("jvm-kotlin")
+    quackquack("kotlin-explicit-api")
     quackquack("test-kotest")
     quackquack("quack-publishing")
+}
+
+dependencies {
+    implementations(
+        libs.kotlin.ksp.api,
+        libs.kotlin.kotlinpoet.core,
+        libs.kotlin.embeddable.compiler,
+    )
 }
