+++ util-backend/src/main/kotlin/team/duckie/quackquack/util/backend/names.kt
@@ -1,0 +1,17 @@
+/*
+ * Designed and developed by Duckie Team 2023.
+ *
+ * Licensed under the MIT.
+ * Please see full license: https://github.com/duckie-team/quack-quack-android/blob/main/LICENSE
+ */
+
+package team.duckie.quackquack.util.backend
+
+import org.jetbrains.kotlin.name.FqName
+
+internal const val QuackComponentPrefix = "Quack"
+
+internal const val ModifierSn = "Modifier"
+internal const val UnitSn = "Unit"
+
+internal val ComposableFqn = FqName("androidx.compose.runtime.Composable")
